Регистрация пользователя в системе (пользователя и администратора) Редактирование профиля.

+++ src/main/java/com/senla/course/announcementPlatform/service/AnnouncementServiceImpl.java
@@ -3,6 +3,8 @@ package com.senla.course.announcementPlatform.service;
 import com.senla.course.announcementPlatform.dao.AnnouncementDao;
 import com.senla.course.announcementPlatform.model.Announcement;
 import com.senla.course.announcementPlatform.service.interfaces.AnnouncementService;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -10,7 +12,7 @@ import java.util.List;
 
 @Service
 public class AnnouncementServiceImpl implements AnnouncementService {
-
+    private static final Logger logger = LogManager.getLogger();
     @Autowired
     private AnnouncementDao announcementDao;
 
@@ -21,7 +23,12 @@ public class AnnouncementServiceImpl implements AnnouncementService {
 
     @Override
     public void delete(Integer id) {
-        announcementDao.delete(getById(id));
+        Announcement announcement = null;
+        if (announcementDao.getById(id) != null) {
+            announcementDao.delete(getById(id));
+        } else {
+            logger.error("Object does not exist");
+        }
     }
 
     @Override
