+++ src/main/java/com/senla/course/announcementPlatform/service/ChatServiceImpl.java
@@ -3,13 +3,15 @@ package com.senla.course.announcementPlatform.service;
 import com.senla.course.announcementPlatform.dao.ChatDao;
 import com.senla.course.announcementPlatform.model.Chat;
 import com.senla.course.announcementPlatform.service.interfaces.ChatService;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 @Service
 public class ChatServiceImpl implements ChatService {
-
+    private static final Logger logger = LogManager.getLogger();
     @Autowired
     private ChatDao chatDao;
 
@@ -20,7 +22,11 @@ public class ChatServiceImpl implements ChatService {
 
     @Override
     public void delete(Integer id) {
-        chatDao.delete(getById(id));
+        if (chatDao.getById(id) != null) {
+            chatDao.delete(getById(id));
+        } else {
+            logger.error("Object does not exist");
+        }
     }
 
     @Override
