+++ src/main/java/com/senla/course/announcementPlatform/service/RatingServiceImpl.java
@@ -3,13 +3,15 @@ package com.senla.course.announcementPlatform.service;
 import com.senla.course.announcementPlatform.dao.RatingDao;
 import com.senla.course.announcementPlatform.model.Rating;
 import com.senla.course.announcementPlatform.service.interfaces.RatingService;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 @Service
 public class RatingServiceImpl implements RatingService {
-
+    private static final Logger logger = LogManager.getLogger();
     @Autowired
     private RatingDao ratingDao;
 
@@ -20,7 +22,11 @@ public class RatingServiceImpl implements RatingService {
 
     @Override
     public void delete(Integer id) {
-        ratingDao.delete(getById(id));
+        if (ratingDao.getById(id) != null) {
+            ratingDao.delete(getById(id));
+        } else {
+            logger.error("Object does not exist");
+        }
     }
 
     @Override
