+++ src/main/java/com/senla/course/announcementPlatform/service/UserServiceImpl.java
@@ -3,6 +3,8 @@ package com.senla.course.announcementPlatform.service;
 import com.senla.course.announcementPlatform.dao.UserDao;
 import com.senla.course.announcementPlatform.model.User;
 import com.senla.course.announcementPlatform.service.interfaces.UserService;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -10,7 +12,7 @@ import java.util.List;
 
 @Service
 public class UserServiceImpl implements UserService {
-
+    private static final Logger logger = LogManager.getLogger();
     @Autowired
     private UserDao userDao;
 
@@ -21,7 +23,11 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void delete(Integer id) {
-        userDao.delete(getById(id));
+        if (userDao.getById(id) != null) {
+            userDao.delete(getById(id));
+        } else {
+            logger.error("Object does not exist");
+        }
     }
 
     @Override
